Organized TS types file
